マークアップ言語関連の命令を分離
