+++ demo/basic.html
@@ -61,5 +61,6 @@
 <!-- editor -->
 <!--React.jsのスクリプトはHTMLの末尾で読み込む-->
 <script src="../release/editor.js"></script>
+<script src="../release/plugin_markup.js"></script>
 </body>
 </html>
