+++ demo/runscript2.html
@@ -5,6 +5,7 @@
   <title>なでしこ3 - script要素を実行するデモ</title>
   <!-- なでしこを使う準備 -->
   <script src="../release/wnako3.js?run"></script>
+  <script src="../release/plugin_markup.js"></script>
   <script type="なでしこ">
     # # # # # # # # # # # # # # # # #
     # なでしこのプログラムをここに書く
