+++ demo/turtle2.html
@@ -6,6 +6,7 @@
   <title>なでしこ3/タートルグラフィックス</title>
   <!-- なでしこを使う準備 -->
   <script src="../release/wnako3.js"></script>
+  <script src="../release/plugin_markup.js"></script>
   <script src="../release/plugin_turtle.js"></script>
   <!-- なでしこを呼び出して使う -->
   <script src="common.js"></script>
@@ -102,7 +103,7 @@
 　　(サイズ/3)と(深度-1)でフラクタル。
 　　60だけカメ左回転。
 　　(サイズ/3)と(深度-1)でフラクタル。
-ここまで。                                             
+ここまで。
 カメ作成。
 1にカメ速度設定。# 最速に
 [10,200]にカメ起点移動。
