+++ demo/turtle3.html
@@ -6,6 +6,7 @@
   <title>なでしこ3/タートルグラフィックス</title>
   <!-- なでしこを使う準備 -->
   <script src="../release/wnako3.js"></script>
+  <script src="../release/plugin_markup.js"></script>
   <script src="../release/plugin_turtle.js"></script>
   <!-- なでしこを呼び出して使う -->
   <script src="common.js"></script>
