+++ src/plugin_system.js
@@ -2131,22 +2131,6 @@ const PluginSystem = {
       const browserslist = require('browserslist')
       return browserslist()
     }
-  },
-  'マークダウンHTML変換': { // @マークダウン形式で記述された文字列SをHTML形式に変換する // @まーくだうんえいちてぃーえむえるへんかん
-    type: 'func',
-    josi: [['を']],
-    fn: function (s) {
-      const markdown = require('markdown')
-      return markdown.parse(s)
-    }
-  },
-  'HTML整形': { // @HTML形式で記述された文字列Sを整形する // @えいちてぃーえむえるせいけい
-    type: 'func',
-    josi: [['を']],
-    fn: function (s) {
-      const html = require('html')
-      return html.prettyPrint(s, {indent_size: 2})
-    }
   }
 }
